another round of adapter refactoring, everything works
