+++ ow_plexil/src/plexil-adapter/OwAdapter.cpp
@@ -24,17 +24,8 @@
 
 // C++
 #include <map>
-#include <mutex>
 using std::string;
 using std::vector;
-
-
-///////////////////////////// Conveniences //////////////////////////////////
-
-// A prettier name for the "unknown" value.
-//static Value const Unknown;
-const Value Unknown;
-
 
 
 //////////////////////// PLEXIL Lookup Support //////////////////////////////
@@ -140,82 +131,6 @@ static bool lookup (const std::string& state_name,
   else retval = false;
 
   return retval;
-}
-
-//////////////////////////// Command Handling //////////////////////////////
-
-static int CommandId = 0;
-
-static std::mutex g_shared_mutex;
-using CommandRecord = std::tuple<Command*,
-                                 AdapterExecInterface*,
-                                 bool>;
-
-enum CommandRecordFields {CR_COMMAND, CR_ADAPTER, CR_ACK_SENT};
-
-static std::map<int, std::unique_ptr<CommandRecord>> CommandRegistry;
-
-static std::unique_ptr<CommandRecord>&
-new_command_record(Command* cmd, AdapterExecInterface* intf)
-{
-  auto cr = std::make_tuple(cmd, intf, false);
-  CommandRegistry[++CommandId] = std::make_unique<CommandRecord>(cr);
-  return CommandRegistry[CommandId];
-}
-
-static void ack_command (Command* cmd,
-                         PLEXIL::CommandHandleValue handle,
-                         AdapterExecInterface* intf)
-{
-  intf->handleCommandAck(cmd, handle);
-  intf->notifyOfExternalEvent();
-}
-
-static void ack_success (Command* cmd, AdapterExecInterface* intf)
-{
-  ack_command (cmd, COMMAND_SUCCESS, intf);
-}
-
-static void ack_failure (Command* cmd, AdapterExecInterface* intf)
-{
-  ack_command (cmd, COMMAND_FAILED, intf);
-}
-
-
-static void ack_sent (Command* cmd, AdapterExecInterface* intf)
-{
-  ack_command (cmd, COMMAND_SENT_TO_SYSTEM, intf);
-}
-
-static void send_ack_once(CommandRecord& cr, bool skip=false)
-{
-  std::lock_guard<std::mutex> g(g_shared_mutex);
-  bool& sent_flag = std::get<CR_ACK_SENT>(cr);
-  if (!sent_flag)
-  {
-    if (!skip) {
-      ack_sent(std::get<CR_COMMAND>(cr), std::get<CR_ADAPTER>(cr));
-    }
-    sent_flag = true;
-  }
-}
-
-static void command_status_callback (int id, bool success)
-{
-  auto it = CommandRegistry.find(id);
-  if (it == CommandRegistry.end())
-  {
-    ROS_ERROR_STREAM("command_status_callback: no command registered under id"
-                     << id);
-    return;
-  }
-
-  std::unique_ptr<CommandRecord>& cr = it->second;
-  Command* cmd = std::get<CR_COMMAND>(*cr);
-  AdapterExecInterface* intf = std::get<CR_ADAPTER>(*cr);
-  send_ack_once(*cr, true);
-  if (success) ack_success (cmd, intf);
-  else ack_failure (cmd, intf);
 }
 
 static void stow (Command* cmd, AdapterExecInterface* intf)
